refactor: update subscription validation and improve code organization
